refactor: move vein generator to generator.veins package
