+++ common/src/main/java/com/gregtechceu/gtceu/api/data/worldgen/VeinGenerators.java
@@ -3,7 +3,7 @@ package com.gregtechceu.gtceu.api.data.worldgen;
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.api.addon.AddonFinder;
 import com.gregtechceu.gtceu.api.addon.IGTAddon;
-import com.gregtechceu.gtceu.api.data.worldgen.generator.*;
+import com.gregtechceu.gtceu.api.data.worldgen.generator.veins.*;
 import com.mojang.serialization.Codec;
 import net.minecraft.resources.ResourceLocation;
